% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysisfunctions.R
\name{add.mcmc}
\alias{add.mcmc}
\title{add.mcmc}
\usage{
add.mcmc(x, y)
}
\arguments{
\item{x}{A list of samples created from the \code{coda.samples} function.}

\item{y}{A list of samples created from the \code{coda.samples} function.}
}
\value{
An \code{mcmc.list} object containing both samples
}
\description{
Combines two objects from \code{coda.samples} and converts the
result to an \code{mcmc.list} object complete with the names of the saved
parameters.
}
