% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphingfunctions.R
\name{errdata}
\alias{errdata}
\title{errdata}
\usage{
errdata(res)
}
\arguments{
\item{res}{A data frame containing ID and grouping variables for each
participant in addition to probabilities of each treatment being
better, worse, or the same as each other treatment under consideration. The
columns of this dataframe should be as follows:
first the ID variable, then any desired grouping variables, and then, in any
order,
         \item[] \code{comp}: the two treatments being compared in the row
         (in the form trt1-trt2)
         \item[] \code{firstbetter}: the probability that the first treatment
         is better than the second treatment
         \item[] \code{secondbetter}: the probability that the second
         treatment is better than the first treatment
         \item[] \code{neitherbetter}: 1- \code{firstbetter}-
         \code{secondbetter}}
}
\value{
Returns a data frame with variables representing the winsorized
medians and confidence intervals as well as a cutoff value representing
the magnitude of the y limits of the plot
}
\description{
Generates a data frame containing winsorized confidence interval
information that can then be passed to the \code{ploterrBar} function
}
