% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysisfunctions.R
\name{find.max.gelman}
\alias{find.max.gelman}
\title{find.max.gelman}
\usage{
find.max.gelman(samples)
}
\arguments{
\item{samples}{A list of samples created from the \code{coda.samples} function.}
}
\value{
Returns the maximum Gelman and Rubin's convergence diagnostic for the
samples.
}
\description{
Calculates the maximum Gelman and Rubin's convergence diagnostic
for a sample generated from \code{coda.samples}.
}
