% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphingfunctions.R
\name{gen.DfRect}
\alias{gen.DfRect}
\title{gen.DfRect}
\usage{
gen.DfRect(res, ninfo)
}
\arguments{
\item{res}{A data frame containing ID and grouping variables for each
participant in addition to probabilities of each treatment being
better, worse, or the same as each other treatment under consideration. The
columns of this dataframe should be as follows:
first the ID variable, then any desired grouping variables, and then, in any
order,
     \begin{itemize}
         \item[] \code{comp}: the two treatments being compared in the row
         (in the form trt1-trt2)
         \item[] \code{firstbetter}: the probability that the first treatment
         is better than the second treatment
         \item[] \code{secondbetter}: the probability that the second
         treatment is better than the first treatment
         \item[] \code{neitherbetter}: 1- \code{firstbetter}-
         \code{secondbetter}
     \end{itemize}}

\item{ninfo}{The number of grouping and stratifying variables
(including the ID variable) included in the \code{res} dataframe}
}
\value{
Returns a data frame with variables representing the breaks and
regions of probability bars that can be plotted with the the
\code{plot.combined.probBar} function.
}
\description{
Generates rectangle probability data frames from a summary
dataset that can then be passed to the \code{plot.probBar} function
}
