% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indiv.summary.R
\name{indiv.summary}
\alias{indiv.summary}
\title{indiv.summary}
\usage{
indiv.summary(samples)
}
\arguments{
\item{samples}{MCMC samples of each parameter}
}
\value{
A list containing a list of kernel density plots for each parameter
and a summary table of the parameter distributions
}
\description{
Creates density plots for each parameter and a summary table
results.table creates an overall data frame of organized results for everyone
Creates kernel density plots and summary tables for each parameter from an
individual analysis
}
