% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jags.fit.R
\name{jags.fit}
\alias{jags.fit}
\title{jags.fit}
\usage{
jags.fit(nof1)
}
\arguments{
\item{nof1}{An nof1 object of class "nof1.data" created from the
\code{\link{nof1.data}} function}
}
\value{
A list with the following elements:
\item{burnin}{Half of the converged sequence is thrown out as a burnin}
\item{n.thin}{If the number of iterations that the user wants (\code{n.run})
is less than the number of converged sequences after burnin, the sequence is
 thinned and the thinning interval is stored}
\item{samples}{MCMC samples stored using jags. The returned samples have the
form of mcmc.list and can be directly applied to coda functions}
\item{max.gelman}{Maximum Gelman and Rubin's convergence diagnostic
calculated for the final sample}
\item{dic}{Deviance information criterion from \code{dic.samples}}
\item{eff}{Minimum effective sample size out of all saved parameters}
}
\description{
The jags.fit function calls the \code{jags.model}, \code{adapt},
and \code{coda.samples} functions from the \code{rjags} package to run the
model and extract samples.
}
