% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysisfunctions.R
\name{new.mcmc}
\alias{new.mcmc}
\title{new.mcmc}
\usage{
new.mcmc(x)
}
\arguments{
\item{x}{A list of samples created from the \code{coda.samples} function.}
}
\value{
An \code{mcmc.list} object containing the samples
}
\description{
Converts an object from \code{coda.samples} to an
\code{mcmc.list} object complete with the names of the saved parameters.
}
