% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nof1function.R
\name{nof1function}
\alias{nof1function}
\title{nof1function}
\usage{
nof1function(
  data,
  ID,
  groupvars,
  stratvars,
  response = "normal",
  Y,
  treatment,
  corr.y,
  bs.trend,
  y.time,
  knots.bt.block,
  block.no,
  bs.df,
  beta.prior = list("dnorm", 0, 1e-06),
  hy.prior = list("dgamma", 0.001, 0.001),
  rho.prior = list("dunif", -1, 1),
  eta.prior = list("dnorm", 0, 1e-06),
  n.chains,
  max.run,
  setsize,
  n.run,
  conv.limit,
  clinicaldiff,
  alpha = 0.05,
  a
)
}
\arguments{
\item{data}{A dataframe containing all of the individual N of 1 data}

\item{ID}{The name of the ID variable in the dataset}

\item{groupvars}{The names of grouping variables that the user would like to
be included in the summary table (not including the stratifying variables)}

\item{stratvars}{The names of variables that the user would like the plots to
 be faceted by. A separate plot will be created for each level of the first
stratifying variable, and these plots will be arranged next to each other. If
a second stratifying variable is specified, this variable will be the facet
on the x axis for each of the plots.}

\item{response}{Type of outcome. Right now, it can only be "normal" for a
continuous outcome}

\item{Y}{The name of the outcome variable in the dataset. This variable
should be a vector with NA's included in time order.}

\item{treatment}{The name of the treatment variable in the dataset. This
variable should have the same length as the outcome variable and can be
character or numeric.}

\item{corr.y}{Indicator for whether the correlation among measurements
should be modeled. The default is \code{F}.}

\item{bs.trend}{Indicator for whether the model should adjust for trend using
splines. The default is \code{F}.}

\item{y.time}{Parameter used for modeling splines (time when the outcome
is measured)}

\item{knots.bt.block}{Parameter used for modeling splines. Indicator for
whether or not knots should be set
at the end of each block (except for the last block). If \code{TRUE}, user
should then specify \code{block.no};
if \code{FALSE}, user should then specify \code{bs.df}.}

\item{block.no}{Block number used for modeling splines for the setting where
the knots are set at the end
of each block. Block number with the same length as the outcome.}

\item{bs.df}{Degrees of freedom for modeling splines when knots are not set
at the end of each block.}

\item{beta.prior}{Prior for the treatment-specific intercept. It should be a
list, where the first element is the distribution and the next two are the
parameters associated with the distribution. For example,
list("dnorm", 0, 1e-6) gives a normal prior with mean 0 and standard
deviation 1e-6. If truncation is desired, the last two parameters should be
the upper and lower limits for the truncation.}

\item{hy.prior}{Prior for the heterogeneity parameter. Supports uniform,
gamma, and half normal. It should also be a list of length 3, where the first
element is the distribution (one of dunif, dgamma, or dhnorm)
and the next two are the parameters associated with the distribution.}

\item{rho.prior}{Prior for the correlated error model. It should be also be
a list of length 3, where the first element is the distribution
and the next two are the parameters associated with the distribution.}

\item{eta.prior}{Prior for modelling spline terms. It should also be a list
of length 3, where first element is the distribution
and the next two are the parameters associated with the distribution.}

\item{n.chains}{Number of chains to run}

\item{max.run}{Maximum number of iterations that the user is willing to run.
If the algorithm is not
converging, it will run up to \code{max.run} iterations before printing a
message that it did not converge}

\item{setsize}{Number of iterations that are run between convergence checks.
If the algorithm converges
quickly, a large \code{setsize} is not needed. The number that is printed
between each convergence check is the
gelman-rubin diagnostic, and the function checks whether this number is below
the user-specified conv.limit.}

\item{n.run}{Final number of iterations that the user wants to store. If the
user wants to store fewer iterations
after the algorithm has converged, the sequence will be thinned. If the user
wants to store more iterations,
extra iterations will be run to reach the specified number of runs.}

\item{conv.limit}{Convergence limit for Gelman and Rubin's convergence
diagnostic.}

\item{clinicaldiff}{The lowest possible practical clinical difference between
treatment coefficients}

\item{alpha}{The desired alpha value for confidence intervals. The default is
0.05.}

\item{a}{A vector of random seeds that is at least as long as the number of
individuals in the dataset}
}
\value{
A list of the following elements:
\item{indivsummary}{A table of summary statistics for each individual,
including their ID, the specified \code{groupvars} and \code{stratvars},
the means and medians of the treatment specific intercepts, the posterior
quantiles of each treatment minus each other treatment, the posterior
probability that each treatment coefficient minus each other is greater than
0, and the probability that each treatment is at least \code{clinicaldiff}
greater than and less than each other
treatment}
\item{indivresults}{A list of the following elements for each individual:
a list of kernel density plots for each parameter for that individual, a
list of kernel density plots for each parameter for that individual, and
the model dic for that individual's analysis}
\item{probplot}{A probability bar plot for each treatment comparison,
stratified by the specified \code{stratvars}}
\item{forestplot}{A winsorized forest error bar plot, stratified by the
specified \code{stratvars}}
}
\description{
This is the wrapper function that takes in a user specified
dataset, variables, priors, etc., runs the individual analyses on each
individual in the dataset, and outputs a summary table, a probability bar
plot, a forest plot, and density plots, dic information, and parameter
summary statistics for each individual.
}
