% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphingfunctions.R
\name{plot.arrange}
\alias{plot.arrange}
\title{plot.arrange}
\usage{
\method{plot}{arrange}(res.ind, groupvars, stratvars, clinicaldiff)
}
\arguments{
\item{res.ind}{A data frame generated from \code{results.table} containing ID
and grouping variables for each participant in addition to probabilities of
each treatment being better, worse, or the same as each other treatment under
consideration. The columns of this dataframe should be as follows:
first the ID variable, then any desired grouping variables, and then, in any
order,
     \begin{itemize}
         \item[] \code{comp}: the two treatments being compared in the row
         (in the form trt1-trt2)
         \item[] \code{firstbetter}: the probability that the first treatment
         is better than the second treatment
         \item[] \code{secondbetter}: the probability that the second
         treatment is better than the first treatment
         \item[] \code{neitherbetter}: 1- \code{firstbetter}-
         \code{secondbetter}
     \end{itemize}}

\item{groupvars}{The names of grouping variables that the user would like to
be included in the summary table (not including the stratifying variables)}

\item{stratvars}{The names of variables that the user would like the plots to
 be faceted by. A separate plot will be created for each level of the first
stratifying variable, and these plots will be arranged next to each other. If
a second stratifying variable is specified, this variable will be the facet
on the x axis for each of the plots.}

\item{clinicaldiff}{The lowest possible practical clinical difference between
treatment coefficients}
}
\value{
Returns a stratified probability bar plot from the
\code{plot.probBar} function and a forest plot from the \code{plot.errBar}
function that can be viewed using \code{grid.draw()}.
}
\description{
Generates stratified, faceted probability bar plots and forest
plots and arranges them nicely
}
