% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphingfunctions.R
\name{plot.errBar}
\alias{plot.errBar}
\title{plot.errBar}
\usage{
\method{plot}{errBar}(
  data,
  title,
  title.size = 10,
  facets,
  yintercept,
  ref.col = "grey80",
  errBar.width = 0.5,
  x.colname = "ID",
  y.colname = c("comp2.5_cut", "comp50_cut", "comp97.5_cut"),
  x.label = "",
  panel.y,
  label.y,
  y.label = "Average difference",
  clinicaldiff,
  cutoff
)
}
\arguments{
\item{data}{A data frame containing the confidence interval values to be
plotted}

\item{title}{The desired title for the plot}

\item{title.size}{The desired title size for the plot}

\item{facets}{The desired facets for the plot}

\item{yintercept}{A vector of the desired lower and upper bounds for the
middle shaded region of the plot}

\item{ref.col}{The desired color of the middle shaded region of the plot}

\item{errBar.width}{The desired width of the error bars}

\item{x.colname}{The name of the variable to be plotted on the x axis
(this is the "ID" variable)}

\item{y.colname}{A vector of the names of the variables to be plotted on the
y axis (these are the lower, middle, and upper confidence interval values to
be plotted)}

\item{x.label}{The desired label for the x axis}

\item{panel.y}{An indicator for whether the y axis facet labels should be
included in the plot}

\item{label.y}{An indicator for whether the y axis label should be included
in the plot}

\item{y.label}{The desired label for the y axis}

\item{clinicaldiff}{The lowest possible practical clinical difference between
treatment coefficients}

\item{cutoff}{Desired magnitude of the y limits of the plot}
}
\value{
Returns a forest error bar plot with the first facet variable as a
y-axis facet, the second facet variable as an x-axis facet (if specified),
and the 95% confidence interval for each \code{x.colname} value.
}
\description{
Creates a forest error bar plot.
}
