% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.table.R
\name{results.table}
\alias{results.table}
\title{results.table}
\usage{
results.table(
  indiv,
  treatment,
  groupvars,
  stratvars,
  data,
  data.ind,
  Y,
  nof1,
  samples,
  clinicaldiff,
  alpha = 0.05
)
}
\arguments{
\item{indiv}{ID of the individual}

\item{treatment}{The name of the treatment variable in the data set. This
variable should have the same length as the outcome variable and can be
character or numeric.}

\item{groupvars}{The names of grouping variables that the user would like to
be included in the summary table (not including the stratifying variables)}

\item{stratvars}{The names of variables that the user would like the plots to
 be faceted by. A separate plot will be created for each level of the first
stratifying variable, and these plots will be arranged next to each other. If
a second stratifying variable is specified, this variable will be the facet
on the x axis for each of the plots.}

\item{data.ind}{A data frame containing all of the individual's N of 1 data}

\item{Y}{The name of the outcome variable in the dataset. This variable
should be a vector with NA's included in time order.}

\item{nof1}{Modeling result of class \code{nof1.data} produced by
\code{nof1.data}}

\item{samples}{MCMC samples of each parameter}

\item{clinicaldiff}{The lowest possible practical clinical difference between
treatment coefficients}

\item{alpha}{The desired alpha value for confidence intervals. The default is
0.05.}
}
\value{
Returns a list of summary statistics of the raw data and the
fitted model for each treatment comparison for the individual.
\item{raw.y.mean}{The raw mean of the outcome for each treatment}
\item{raw.y.median}{The raw median of the outcome for each treatment}
\item{post.coef.mean}{The posterior mean of the coefficient for each
treatment}
\item{post.coef.median}{The posterior median of the coefficient for each
treatment}
\item{post.y.mean}{The posterior mean of the outcome for each treatment}
\item{post.y.median}{The posterior median of the outcome for each treatment}
\item{post.coef.ci}{The credible interval of the coefficient for each
treatment}
\item{post.y.ci}{The credible interval of the outcome for each treatment}
\item{comp.treat.post.coef}{The posterior quantiles of one coefficient minus
the other when comparing two treatments}
\item{p.comp.coef.greater.0}{The posterior probability of one coefficient
minus the other is greater than 0}
\item{comp.treat.post.y}{The posterior quantiles of one outcome minus the
other when comparing two treatments}
}
\description{
Computes and presents a summary of the results from an
individual analysis.
}
