% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nof1.inits.normal.R
\name{nof1.inits.normal}
\alias{nof1.inits.normal}
\title{nof1.inits.normal}
\usage{
nof1.inits.normal(data, Treat.name, bs.trend, bs_df, n.chains)
}
\arguments{
\item{data}{A list containing Y (the outcome), vectors of 0s and 1s in
the model matrix for each \emph{Treat.name}, indicating whether or not the
response was recorded under the treatment, and base spline information}
}
\value{
A list of the initial values for each treatment-specific intercept
and spline term to be used in the \code{jags.fit} function
}
\description{
The nof1.inits.normal function generates initial parameter
values to be passed to the \code{jags.fit} function.
}
